+++ common/map.h
@@ -470,7 +470,7 @@ extern struct tile_type tile_types[T_LAST];
  */
 #define circle_iterate(center_x, center_y, sq_radius, x_itr, y_itr)           \
 {                                                                             \
-  int _cr_radius = (int)sqrt(sq_radius);                                     \
+  int _cr_radius = (int)sqrt((double)(sq_radius));                            \
   square_dxy_iterate(center_x, center_y, _cr_radius,                          \
 		     x_itr, y_itr, _dx, _dy) {                                \
     if (_dy * _dy + _dx * _dx <= (sq_radius)) {
